+++ cmd/chkacehealthy/main.go
@@ -18,17 +18,97 @@ limitations under the License.
 package main
 
 import (
+	"context"
 	"fmt"
+	"log"
 	"net"
+	"net/http"
 	"os"
 	"time"
+
+	"k8s.io/apimachinery/pkg/api/errors"
 )
+
+var checkACE = checkACElocal
+var httpCheck = httpChecklocal
+var socketCheck = socketChecklocal
+var osExit = os.Exit
 
 const restartIsTimeoutInSeconds = 60
 
+var netDial = net.Dial
+var httpGet = http.Get
+
 func main() {
+
+	err := checkACE()
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	// If knative service also check FDR is up
+	knative := os.Getenv("KNATIVESERVICE")
+	if knative == "true" || knative == "1" {
+		fmt.Println("KNATIVESERVICE set so checking FDR container")
+		err := checkDesignerHealth()
+		if err != nil {
+			log.Fatal(err)
+		}
+	} else {
+		fmt.Println("KNATIVESERVICE is not set so skipping FDR checks")
+	}
+
+}
+
+func checkDesignerHealth() error {
+	// HTTP LMAP endpoint
+	err := httpCheck("LMAP Port", "http://localhost:3002/admin/ready")
+	if err != nil {
+		return err
+	}
+
+	isConnectorService := os.Getenv("CONNECTOR_SERVICE")
+	if isConnectorService == "true" || isConnectorService == "1" {
+		// HTTP LCP Connector service endpoint
+		err = httpCheck("LCP Port", "http://localhost:3001/admin/ready")
+		if err != nil {
+			return err
+		}
+	}
+
+	// LCP api flow endpoint
+	lcpsocket := "/tmp/lcp.socket"
+	if value, ok := os.LookupEnv("LCP_IPC_PATH"); ok {
+		lcpsocket = value
+	}
+	err = socketCheck("LCP socket", lcpsocket)
+	if err != nil {
+		return err
+	}
+
+	// LMAP endpoint
+	lmapsocket := "/tmp/lmap.socket"
+	if value, ok := os.LookupEnv("LMAP_IPC_PATH"); ok {
+		lmapsocket = value
+	}
+	err = socketCheck("LMAP socket", lmapsocket)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func isEnvExist(key string) bool {
+	if _, ok := os.LookupEnv(key); ok {
+		return true
+	}
+	return false
+}
+
+func checkACElocal() error {
 	// Check if the integration server has started the admin REST endpoint
-	conn, err := net.Dial("tcp", "127.0.0.1:7600")
+	conn, err := netDial("tcp", "127.0.0.1:7600")
 
 	if err != nil {
 
@@ -38,10 +118,10 @@ func main() {
 
 		if os.IsNotExist(statErr) {
 			fmt.Println("Integration server is not active")
-			os.Exit(1)
-		} else if statErr != nil  {
+			return errors.NewBadRequest("Integration server is not active")
+		} else if statErr != nil {
 			fmt.Println(statErr)
-			os.Exit(1)
+			return errors.NewBadRequest("stat error " + statErr.Error())
 		} else {
 			fmt.Println("Integration server restart file found")
 			timeNow := time.Now()
@@ -49,13 +129,49 @@ func main() {
 
 			if timeDiff.Seconds() < restartIsTimeoutInSeconds {
 				fmt.Println("Integration server is restarting")
-				os.Exit(0)
 			} else {
 				fmt.Println("Integration restart time elapsed")
-				os.Exit(1)
+				return errors.NewBadRequest("Integration restart time elapsed")
 			}
 		}
+	} else {
+		fmt.Println("ACE ready check passed")
 	}
 	conn.Close()
+	return nil
+}
 
+func httpChecklocal(name string, addr string) error {
+	resp, err := httpGet(addr)
+	if err != nil {
+		return err
+	}
+	if resp.StatusCode != 200 {
+		fmt.Println(name + " ready check failed - HTTP Status is not 200 range")
+		return errors.NewBadRequest(name + " ready check failed - HTTP Status is not 200 range")
+	} else {
+		fmt.Println(name + " ready check passed")
+	}
+	return nil
+}
+
+func socketChecklocal(name string, socket string) error {
+	httpc := http.Client{
+		Transport: &http.Transport{
+			DialContext: func(_ context.Context, _, _ string) (net.Conn, error) {
+				return net.Dial("unix", socket)
+			},
+		},
+	}
+	response, err := httpc.Get("http://dummyHostname/admin/ready")
+	if err != nil {
+		return err
+	}
+	if response.StatusCode != 200 {
+		log.Fatal(name + " ready check failed - HTTP Status is not 200 range")
+		return errors.NewBadRequest(name + " ready check failed - HTTP Status is not 200 range")
+	} else {
+		fmt.Println(name + " ready check passed")
+	}
+	return nil
 }